:recycle: REFACTOR: 자동 계산 로직 자동 계산 서비스 클래스로 이동
