+++ slash/src/main/java/project/slash/statistics/controller/IncidentTaskController.java
@@ -8,12 +8,12 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import project.slash.common.response.BaseResponse;
 import project.slash.statistics.dto.request.RequestStatisticsDto;
-import project.slash.statistics.service.StatisticsService;
+import project.slash.statistics.service.AutoStatisticsService;
 
 @RestController
 @RequiredArgsConstructor
 public class IncidentTaskController {
-	private final StatisticsService statisticsService;
+	private final AutoStatisticsService autoStatisticsService;
 
 	/**
 	 * 지정된 날의 해당 월의 1일부터 지정된 날까지 장애 적기 처리율을 저장하는 메서드 입니다.
@@ -23,7 +23,7 @@ public class IncidentTaskController {
 	@PostMapping("/common/incident-statistics")
 	public BaseResponse<Void> addIncidentDueOnTimeRate(
 		@RequestBody @Valid RequestStatisticsDto requestStatisticsDto) {
-		statisticsService.getIncidentStatistics(requestStatisticsDto);
+		autoStatisticsService.getIncidentStatistics(requestStatisticsDto);
 		return BaseResponse.ok();
 	}
 }
