+++ slash/src/main/java/project/slash/statistics/controller/StatisticsController.java
@@ -16,9 +16,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import lombok.RequiredArgsConstructor;
 import project.slash.common.response.BaseResponse;
-import project.slash.statistics.dto.request.RequestStatisticsDto;
 import project.slash.statistics.dto.response.MonthlyServiceStatisticsDto;
-import project.slash.statistics.dto.response.MonthlyStatisticsDto;
 import project.slash.statistics.dto.request.EditStatisticsDto;
 import project.slash.statistics.dto.response.MonthlyIndicatorsDto;
 import project.slash.statistics.dto.response.StatisticsStatusDto;
